Bug fix aft tank volume computation

+++ RCAIDE/Library/Methods/Performance/compute_payload_range_diagram.py
@@ -104,13 +104,12 @@ def compute_payload_range_diagram(mission = None, cruise_segment_tag = "cruise",
     geometry(mission)
     for segment in  mission.segments:
         # perform inital weights analysis
-        segment.analyses.geometry.vehicle.mass_properties.takeoff = None
+        segment.analyses.vehicle.mass_properties.takeoff = None
     mass_properties(mission)
-    vehicle = mission.segments[initial_segment].analyses.weights.vehicle
-    
-
+    vehicle = mission.segments[initial_segment].analyses.vehicle 
   
     [setattr(seg.analyses.aerodynamics.settings, "store_training_data", True) for seg in mission.segments]
+    file_name = [(seg.analyses.aerodynamics.tag) for seg in mission.segments][0]
     for network in vehicle.networks:
         if type(network) == RCAIDE.Framework.Networks.Fuel:  
             payload_range  =  conventional_payload_range_diagram(vehicle,mission,cruise_segment_tag,fuel_reserve_percentage,plot_diagram,fuel_name) 
@@ -119,7 +118,7 @@ def compute_payload_range_diagram(mission = None, cruise_segment_tag = "cruise",
     
     if delete_training_data:
         for fname in os.listdir(os.path.dirname(os.path.abspath(sys.argv[0]))):
-            if fname.endswith(".pkl") and "payload_range_mission" in fname:
+            if fname.endswith(".pkl") and "payload_range_mission" in fname and file_name in fname:
                 os.remove(os.path.join(os.path.dirname(os.path.abspath(sys.argv[0])), fname))
 
     print("\n============== Payload Range Report ==============\n")            
@@ -222,9 +221,9 @@ def conventional_payload_range_diagram(vehicle,mission,cruise_segment_tag,fuel_r
     # loop for each point of Payload Range Diagram
     for i in range(len(TOW)):
         # Define takeoff weight
-        mission.segments[0].analyses.geometry.vehicle.mass_properties.takeoff  = TOW[i]
-        mission.segments[0].analyses.geometry.vehicle.mass_properties.payload  = PLD[i]
-        mission.segments[0].analyses.geometry.vehicle.mass_properties.fuel     = FUEL[i]
+        mission.segments[0].analyses.vehicle.mass_properties.takeoff  = TOW[i]
+        mission.segments[0].analyses.vehicle.mass_properties.payload  = PLD[i]
+        mission.segments[0].analyses.vehicle.mass_properties.fuel     = FUEL[i]
 
 
         # Evaluate mission with current TOW
@@ -371,7 +370,7 @@ def electric_payload_range_diagram(vehicle,mission,cruise_segment_tag,plot_diagr
 
     # Calculate Vehicle Range for Max Payload and Ferry Conditions
     for i in range(2):
-        mission.segments[0].analyses.weights.vehicle.mass_properties.takeoff = TOW[i]
+        mission.segments[0].analyses.vehicle.mass_properties.takeoff = TOW[i]
         results = mission.evaluate()
         segment = results.segments[cruise_segment_tag]
         R[i]    = segment.conditions.frames.inertial.position_vector[-1,0]
